use_cases/RAG/05_Data-Query/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pinecone-workshop-data-query"
version = "0.1.0"
description = "Backend data query pipeline for the Pinecone workshop"
authors = ["Vamshi Enabothala"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9,<3.13"
pinecone-client = "^4.1.1"
python-dotenv = "1.0.1"
requests = "2.31.0"
boto3 = "^1.34.128"
fastapi="^0.110.1"
uvicorn="^0.27.1"
flax="^0.8.2"
pillow="^10.3.0"

[tool.poetry.dev-dependencies]
